Refactor: Rename app from 'nhacte' to 'nhac'
- Updated app name across all platforms (Android, iOS, Linux, Windows, macOS)
- Changed package name from dev.myyc.nhacte to dev.myyc.nhac
- Updated all class names (NhacteApp -> NhacApp, NhacteAudioHandler -> NhacAudioHandler)
- Modified database name from nhacte_cache.db to nhac_cache.db
- Updated all references in configuration files and build scripts
- Moved Android MainActivity to new package directory structure

This is a comprehensive rename that affects the entire codebase and all platform
configurations.
