+++ android/app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE_MEDIA_PLAYBACK"/>
     
     <application
-        android:label="nhacte"
+        android:label="nhac"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
