+++ android/app/src/main/kotlin/dev/myyc/nhac/MainActivity.kt
@@ -1,4 +1,4 @@
-package dev.myyc.nhacte
+package dev.myyc.nhac
 
 import com.ryanheise.audioservice.AudioServiceActivity
