+++ lib/screens/artists_screen.dart
@@ -94,7 +94,7 @@ class _ArtistsScreenState extends State<ArtistsScreen> {
                       child: CachedNetworkImage(
                         imageUrl: api.getCoverArtUrl(artist.coverArt, size: 100),
                         httpHeaders: const {
-                          'User-Agent': 'nhacte/1.0.0',
+                          'User-Agent': 'nhac/1.0.0',
                         },
                         fit: BoxFit.cover,
                         width: 40,
