+++ macos/Runner/Configs/AppInfo.xcconfig
@@ -5,10 +5,10 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = nhacte
+PRODUCT_NAME = nhac
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = dev.myyc.nhacte
+PRODUCT_BUNDLE_IDENTIFIER = dev.myyc.nhac
 
 // The copyright displayed in application information
 PRODUCT_COPYRIGHT = Copyright © 2025 dev.myyc. All rights reserved.
